pyproject.toml
[tool.poetry]
name = "discord-chatgpt-bot-python"
version = "0.1.0"
description = ""
authors = ["koiking213 <koiking213@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.4"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
